refactor(app-web): move Browsers live-session list into the doc sidebar
The 5th operator app (Browsers) now hangs its live-session list off the
persistent left DocSidebar (BrowsersSidebarPanel), the same slot Tasks /
CRM / Brain use, instead of wedging a second rail into the content pane.
The Take-Over live view fills the full-width pane; the "My Browser"
connect control stays in the surface top bar and the live-session count
moves to the sidebar panel header.
