+++ apps/app-web/src/app/w/[workspaceId]/computer/layout.tsx
@@ -2,9 +2,10 @@
 
 /**
  * Browsers surface layout — every `/w/[id]/computer/*` route renders inside
- * the `BrowsersSurfaceShell` (operator top bar + live-session rail). The rail
- * persists across `/computer/<sessionId>` transitions, so navigating between
- * sessions never remounts the chrome. See computer-use.md §5.
+ * the `BrowsersSurfaceShell` (operator top bar over the full-width live view).
+ * The live-session list lives in the persistent left sidebar
+ * (`BrowsersSidebarPanel`), not this pane, so navigating between sessions
+ * never remounts the chrome. See computer-use.md §5.
  *
  * [COMP:app-web/browsers-surface]
  */
