+++ apps/app-web/src/components/doc/doc-sidebar.tsx
@@ -127,6 +127,7 @@ import { WorkflowSidebarPanel } from "./sidebar-panels/workflow-sidebar-panel";
 import { FeedSidebarPanel } from "./sidebar-panels/feed-sidebar-panel";
 import { TasksSidebarPanel } from "./sidebar-panels/tasks-sidebar-panel";
 import { CrmSidebarPanel } from "./sidebar-panels/crm-sidebar-panel";
+import { BrowsersSidebarPanel } from "./sidebar-panels/browsers-sidebar-panel";
 
 export type SidebarMove = {
   viewId: string;
@@ -721,9 +722,9 @@ export function DocSidebar(props: Props) {
 
       <div className="flex-1 min-h-0 overflow-y-auto px-2 pb-4">
         {/* Surface-aware body. The page tree (Favorites / Drafts / search) shows
-            ONLY on Home (`'p'`); Brain / Studio / Workflow swap in their own
-            controls; every other surface (approvals, knowledge-base, root)
-            renders nothing here. */}
+            ONLY on Home (`'p'`); Brain / Studio / Workflow / Tasks / CRM /
+            Browsers each swap in their own panel; every other surface
+            (approvals, knowledge-base, root) renders nothing here. */}
         {props.activeSurface === "brain" ? (
           <BrainSidebarPanel workspaceId={workspaceId} />
         ) : props.activeSurface === "studio" ? (
@@ -736,6 +737,8 @@ export function DocSidebar(props: Props) {
           <TasksSidebarPanel workspaceId={workspaceId} />
         ) : props.activeSurface === "crm" ? (
           <CrmSidebarPanel workspaceId={workspaceId} />
+        ) : props.activeSurface === "computer" ? (
+          <BrowsersSidebarPanel workspaceId={workspaceId} />
         ) : null}
 
         {props.activeSurface === "p" && (
